selection history no longer depends on blueprint
